Added indel handeling

+++ starka/src/lib/starling_common/gvcf_locus_info.hh
@@ -327,7 +327,7 @@ struct site_info {
     double ReadPosRankSum;   // Uses Mann-Whitney Rank Sum Test for the distance from the end of the read containing an alternate allele.
     double BaseQRankSum;     // Uses Mann-Whitney Rank Sum Test for BQs (ref bases vs alternate alleles)
     double MQRankSum;        // Uses Mann-Whitney Rank Sum Test for MQs (ref bases vs alternate alleles)
-    double Qscore;           // The empirically calibrated quality-score of the site, if -1 not q-score has been reported
+    int Qscore;           // The empirically calibrated quality-score of the site, if -1 not q-score has been reported
 
     site_modifiers smod;
 };
